reserve uT-* + add UNIT_TESTS (was -DNOMAIN) to config.h
I realized "whether we are unit-testing or not" really works better as
part of the configuration.  This removes a need to assign to CPPFLAGS
that might otherwise make life difficult later.

N.B. Even though unit test sources need to show up directly in build
directories, they are NOT meant to live in this repository:(we want to
be able to do checkouts and switch between branches without having the
tests change out from under us, in order to, e.g., verify that
something *was* a bug and that a particular change fixed it), hence
the reservation of the uT-* prefix so that we can have these files
appear as needed but be invisible to (this) git repository.

Also, NOMAIN (as in '#ifndef NOMAIN') was violating
my 'double-negative' rule, so that is deprecated.